Fix loadCats() being called every time the activity is recreated
Calling loadCats() inside the activity onCreate() is problematic because every time the activity is recreated e.g. screen rotation a new API call is made. This removes the purpose of ViewModel to retain informations in lifecycle events.
To fix this, init{} is a good place to make this call as it will only happen on viewModel creation.